requirements.txt
Django==1.10.1
django-bootstrap3-datetimepicker==2.2.3
django-widget-tweaks==1.4.1
djangorestframework==3.4.7
wheel==0.24.0
